+++ bootique-jersey-client/src/main/java/io/bootique/jersey/client/HttpClientFactory.java
@@ -1,31 +1,48 @@
 package io.bootique.jersey.client;
 
-import javax.ws.rs.client.Client;
-
 import io.bootique.jersey.client.auth.AuthenticatorFactory;
 
+import javax.ws.rs.client.Client;
+
+/**
+ * An injectable factory for creating JAX RS clients based on Bootique configuration.
+ */
 public interface HttpClientFactory {
 
-	/**
-	 * Returns a new instance of JAX-RS HTTP client initialized using
-	 * "jerseyclient" configuration subtree.
-	 * 
-	 * @return a new instance of JAX-RS HTTP client initialized using
-	 *         "jerseyclient" configuration subtree.
-	 */
-	Client newClient();
+    /**
+     * Returns a new instance of JAX-RS HTTP client initialized using
+     * "jerseyclient" configuration subtree.
+     *
+     * @return a new instance of JAX-RS HTTP client initialized using
+     * "jerseyclient" configuration subtree.
+     */
+    default Client newClient() {
+        return newBuilder().build();
+    }
 
-	/**
-	 * Returns a new instance of JAX-RS HTTP client initialized using
-	 * "jerseyclient" configuration subtree and associated with named
-	 * authenticator.
-	 * 
-	 * @return a new instance of JAX-RS HTTP client initialized using
-	 *         "jerseyclient" configuration subtree and associated with named
-	 *         authenticator.
-	 * 
-	 * @since 0.2
-	 * @see AuthenticatorFactory
-	 */
-	Client newAuthenticatedClient(String authName);
+    /**
+     * A builder for a new client. Allows to create a client with Bootique configuration-driven settings and select
+     * a preconfigured authentication and trust store.
+     *
+     * @return a builder for a new client.
+     * @since 0.25
+     */
+    HttpClientBuilder newBuilder();
+
+    /**
+     * Returns a new instance of JAX-RS HTTP client initialized using
+     * "jerseyclient" configuration subtree and associated with named
+     * authenticator.
+     *
+     * @return a new instance of JAX-RS HTTP client initialized using
+     * "jerseyclient" configuration subtree and associated with named
+     * authenticator.
+     * @see AuthenticatorFactory
+     * @since 0.2
+     * @deprecated since 0.25 in favor of "newClientBuilder().auth(authName).build()"
+     */
+    @Deprecated
+    default Client newAuthenticatedClient(String authName) {
+        return newBuilder().auth(authName).build();
+    }
 }