+++ bootique-jersey-client/src/test/java/io/bootique/jersey/client/JerseyClientModuleProvider_MetadataIT.java
@@ -85,6 +85,7 @@ public class JerseyClientModuleProvider_MetadataIT {
                 "[followRedirects:boolean]" +
                 "[readTimeoutMs:int]" +
                 "[trustStore:io.bootique.resource.ResourceFactory]" +
-                "[trustStorePassword:java.lang.String]", result);
+                "[trustStorePassword:java.lang.String]" +
+                "[map:trustStores<java.lang.String,io.bootique.jersey.client.TrustStoreFactory>]", result);
     }
 }